Added a new resource
Hi, I have added a new resource for further reading for merge sort. Found this online while looking for some resources to learn sorting algorithms. Topics were explained in detail and found it to be useful and thought of adding it here. Hope this small contribution will help other learners.